refactor: chat pareller route 수정
