+++ src/app/(auth)/(navigationsBarLayout)/chat/[chatRoomId]/page.tsx
